Pyprocessor/requirements.txt
Flask==2.2.2
flask-socketio==5.3.4
eventlet==0.33.0
gunicorn==20.1.0
Werkzeug==2.2.3
requests==2.28.1
numpy>=1.26.0,<2.1.0
opencv-python-headless==4.7.0.72
mediapipe==0.10.18
tensorflow==2.18.0
torch==2.1.0
transformers==4.41.2
peft==0.5.0
accelerate==0.22.0
bitsandbytes==0.40.2
azure-storage-queue
azure-storage-blob
azure-data-tables
httpx==0.27.2
tenacity==8.2.2
dnspython==2.1.0